+++ src/components/burger-ingredients/ingredients-tabs.js
@@ -1,0 +1,27 @@
+import styles from './burger-ingredients.module.css';
+import {Tab} from "@ya.praktikum/react-developer-burger-ui-components";
+import React, {forwardRef} from "react";
+
+const IngredientsTabs = forwardRef(({current, setCurrent}, ref) => {
+  const onTabClick = (tab) => {
+    setCurrent(tab);
+
+    ref.current[tab].clickRef.current.scrollIntoView({behavior: 'smooth'});
+  }
+
+  return (
+    <div className={`${styles.ingredientsTab} mt-5`}>
+      <Tab value='Булки' active={current === 'Булки'} onClick={onTabClick}>
+        Булки
+      </Tab>
+      <Tab value='Соусы' active={current === 'Соусы'} onClick={onTabClick}>
+        Соусы
+      </Tab>
+      <Tab value='Начинки' active={current === 'Начинки'} onClick={onTabClick}>
+        Начинки
+      </Tab>
+    </div>
+  );
+});
+
+export default IngredientsTabs;
